Fix more refactor leftovers, test against actual ID
