+++ docs/logging.asciidoc
@@ -8,10 +8,11 @@ as well as http://www.structlog.org/en/stable/[`structlog`].
 ==== `logging`
 
 We provide a https://docs.python.org/3/library/logging.html#filter-objects[filter]
-which will add two new attributes to any filtered `LogRecord`:
+which will add three new attributes to any filtered `LogRecord`:
 
 * `elasticapm_transaction_id`
-* `elasticapm_transaction_trace_parent_id`
+* `elasticapm_trace_id`
+* `ealsticapm_span_id`
 
 Note that because https://docs.python.org/3/library/logging.html#filter-objects[filters
 are not propagated to descendent loggers], you should add the filter to each of
@@ -33,16 +34,17 @@ logging.getLogger("").addHandler(console)
 ==== `structlog`
 
 We provide a http://www.structlog.org/en/stable/processors.html[processor] for
-http://www.structlog.org/en/stable/[`structlog`] which will add two new keys
+http://www.structlog.org/en/stable/[`structlog`] which will add three new keys
 to the event_dict of any processed event:
 
-* `"elasticapm_transaction_id"`
-* `"elasticapm_transaction_trace_parent_id"`
+* `"transaction.id"`
+* `"trace.id"`
+* `"span.id"`
 
 [source,python]
 ----
 from structlog import PrintLogger, wrap_logger
-from elasticapm.handlers.logging import structlog_processor
+from elasticapm.handlers.structlog import structlog_processor
 
 wrapped_logger = PrintLogger()
 logger = wrap_logger(wrapped_logger, processors=[structlog_processor])
